@@ -18,22 +18,22 @@
     "contentlayer2": "^0.5.0",
     "date-fns": "^2.30.0",
     "motion": "^11.16.0",
-    "next": "^14.2.5",
+    "next": "15.1.4",
     "next-contentlayer2": "^0.5.3",
     "next-themes": "^0.4.4",
-    "react": "^18.3.1",
-    "react-dom": "^18.3.1"
+    "react": "19.0.0",
+    "react-dom": "19.0.0"
   },
   "devDependencies": {
     "@tailwindcss/typography": "^0.5.10",
     "@types/node": "22.0.2",
-    "@types/react": "18.2.61",
-    "@types/react-dom": "18.3.0",
+    "@types/react": "19.0.7",
+    "@types/react-dom": "19.0.3",
     "@typescript-eslint/eslint-plugin": "^6.9.0",
     "@typescript-eslint/parser": "^6.20.0",
     "autoprefixer": "^10.4.19",
     "eslint": "8.56.0",
-    "eslint-config-next": "^14.2.5",
+    "eslint-config-next": "15.1.4",
     "eslint-config-prettier": "^9.1.0",
     "eslint-plugin-tailwindcss": "^3.17.4",
     "postcss": "^8.4.38",
@@ -45,5 +45,9 @@
     "shiki": "^0.14.5",
     "tailwindcss": "^3.4.3",
     "typescript": "^5.7.2"
+  },
+  "overrides": {
+    "@types/react": "19.0.7",
+    "@types/react-dom": "19.0.3"
   }
 }
